+++ emerlin2caom2/file_metadata.py
